+++ src/dma/traits.rs
@@ -262,6 +262,10 @@ impl RccEnable for pac::DMA1 {
             let rcc = &(*RCC::ptr());
             // Enable and reset the timer peripheral
             bb::set(&rcc.ahb1enr, 21);
+
+            // Stall the pipeline to work around erratum 2.1.13 (DM00037591)
+            cortex_m::asm::dsb();
+
             bb::set(&rcc.ahb1rstr, 21);
             bb::clear(&rcc.ahb1rstr, 21);
         }
@@ -275,6 +279,10 @@ impl RccEnable for pac::DMA2 {
             let rcc = &(*RCC::ptr());
             // Enable and reset the timer peripheral
             bb::set(&rcc.ahb1enr, 22);
+
+            // Stall the pipeline to work around erratum 2.1.13 (DM00037591)
+            cortex_m::asm::dsb();
+
             bb::set(&rcc.ahb1rstr, 22);
             bb::clear(&rcc.ahb1rstr, 22);
         }
@@ -409,7 +417,7 @@ address!(
     (pac::I2C3, dr, u8),
 );
 
-#[cfg(not(any(feature = "stm32f410", feature = "stm32f446")))]
+#[cfg(not(any(feature = "stm32f410")))]
 dma_map!(
     (Stream3<DMA2>, Channel4, pac::SDIO, MemoryToPeripheral), //SDIO
     (Stream3<DMA2>, Channel4, pac::SDIO, PeripheralToMemory), //SDIO
@@ -417,19 +425,8 @@ dma_map!(
     (Stream6<DMA2>, Channel4, pac::SDIO, PeripheralToMemory), //SDIO
 );
 
-#[cfg(not(any(feature = "stm32f410", feature = "stm32f446")))]
+#[cfg(not(any(feature = "stm32f410")))]
 address!((pac::SDIO, fifo, u32),);
-
-#[cfg(feature = "stm32f446")]
-dma_map!(
-    (Stream3<DMA2>, Channel4, pac::SDMMC, MemoryToPeripheral), //SDMMC
-    (Stream3<DMA2>, Channel4, pac::SDMMC, PeripheralToMemory), //SDMMC
-    (Stream6<DMA2>, Channel4, pac::SDMMC, MemoryToPeripheral), //SDMMC
-    (Stream6<DMA2>, Channel4, pac::SDMMC, PeripheralToMemory), //SDMMC
-);
-
-#[cfg(feature = "stm32f446")]
-address!((pac::SDMMC, sdmmc_fifor, u32),);
 
 #[cfg(any(
     feature = "stm32f401",