+++ src/spi.rs
@@ -750,6 +750,9 @@ impl<PINS> Spi<SPI1, PINS> {
 
             // Enable clock.
             bb::set(&rcc.apb2enr, EN_BIT);
+
+            // Stall the pipeline to work around erratum 2.1.13 (DM00037591)
+            cortex_m::asm::dsb();
         }
 
         Spi { spi, pins }.init(mode, freq, clocks.pclk2())
@@ -787,6 +790,9 @@ impl<PINS> Spi<SPI2, PINS> {
 
             // Enable clock.
             bb::set(&rcc.apb1enr, EN_BIT);
+
+            // Stall the pipeline to work around erratum 2.1.13 (DM00037591)
+            cortex_m::asm::dsb();
         }
 
         Spi { spi, pins }.init(mode, freq, clocks.pclk1())
@@ -823,6 +829,9 @@ impl<PINS> Spi<SPI3, PINS> {
 
             // Enable clock.
             bb::set(&rcc.apb1enr, EN_BIT);
+
+            // Stall the pipeline to work around erratum 2.1.13 (DM00037591)
+            cortex_m::asm::dsb();
         }
 
         Spi { spi, pins }.init(mode, freq, clocks.pclk1())
@@ -855,6 +864,9 @@ impl<PINS> Spi<SPI4, PINS> {
 
             // Enable clock.
             bb::set(&rcc.apb2enr, EN_BIT);
+
+            // Stall the pipeline to work around erratum 2.1.13 (DM00037591)
+            cortex_m::asm::dsb();
         }
 
         Spi { spi, pins }.init(mode, freq, clocks.pclk2())
@@ -886,6 +898,9 @@ impl<PINS> Spi<SPI5, PINS> {
 
             // Enable clock.
             bb::set(&rcc.apb2enr, EN_BIT);
+
+            // Stall the pipeline to work around erratum 2.1.13 (DM00037591)
+            cortex_m::asm::dsb();
         }
 
         Spi { spi, pins }.init(mode, freq, clocks.pclk2())
@@ -912,6 +927,9 @@ impl<PINS> Spi<SPI6, PINS> {
 
             // Enable clock.
             bb::set(&rcc.apb2enr, EN_BIT);
+
+            // Stall the pipeline to work around erratum 2.1.13 (DM00037591)
+            cortex_m::asm::dsb();
         }
 
         Spi { spi, pins }.init(mode, freq, clocks.pclk2())
@@ -1054,10 +1072,19 @@ where
         let sr = self.spi.sr.read();
 
         Err(if sr.ovr().bit_is_set() {
+            // Read from the DR to clear the OVR bit
+            let _ = self.spi.dr.read();
             nb::Error::Other(Error::Overrun)
         } else if sr.modf().bit_is_set() {
+            // Write to CR1 to clear MODF
+            self.spi.cr1.modify(|_r, w| w);
             nb::Error::Other(Error::ModeFault)
         } else if sr.crcerr().bit_is_set() {
+            // Clear the CRCERR bit
+            self.spi.sr.modify(|_r, w| {
+                w.crcerr().clear_bit();
+                w
+            });
             nb::Error::Other(Error::Crc)
         } else if sr.txe().bit_is_set() {
             // NOTE(write_volatile) see note above