+++ src/timer.rs
@@ -235,6 +235,10 @@ macro_rules! hal {
                         let rcc = &(*RCC::ptr());
                         // Enable and reset the timer peripheral, it's the same bit position for both registers
                         bb::set(&rcc.$apbenr, $en_bit);
+
+                        // Stall the pipeline to work around erratum 2.1.13 (DM00037591)
+                        cortex_m::asm::dsb();
+
                         bb::set(&rcc.$apbrstr, $reset_bit);
                         bb::clear(&rcc.$apbrstr, $reset_bit);
                     }